Fix unit failing due to refactoring
