+++ app/src/main/java/ch/epfl/favo/auth/SignInActivity.java
@@ -2,13 +2,13 @@ package ch.epfl.favo.auth;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.StringRes;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.firebase.ui.auth.AuthUI;
+import com.firebase.ui.auth.ErrorCodes;
 import com.firebase.ui.auth.IdpResponse;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GoogleApiAvailability;
@@ -16,7 +16,6 @@ import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.auth.ActionCodeSettings;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.iid.FirebaseInstanceId;
 
 import java.util.Arrays;
 import java.util.List;
@@ -130,9 +129,6 @@ public class SignInActivity extends AppCompatActivity {
   private void handleSignInResponse(IdpResponse idpResponse, int resultCode) {
 
     if (resultCode == RESULT_OK) {
-
-      retrieveCurrentRegistrationToken();
-
       // Successfully signed in
       startMainActivity();
     } else {
@@ -143,23 +139,6 @@ public class SignInActivity extends AppCompatActivity {
       }
       showSnackbar(R.string.unknown_error);
     }
-  }
-
-  // retrieve current registration token for the notification system
-  private void retrieveCurrentRegistrationToken() {
-    FirebaseInstanceId.getInstance()
-        .getInstanceId()
-        .addOnCompleteListener(
-            task -> {
-              if (!task.isSuccessful()) {
-                return;
-              }
-
-              // Get new Instance ID token
-              String token = Objects.requireNonNull(task.getResult()).getToken();
-              Log.d("SignInActivity", getString(R.string.msg_token_fmt, token));
-              // TODO send registration token to db
-            });
   }
 
   public void showSnackbar(@StringRes int errorMessageRes) {
